rename features modules
