+++ tests/feature/test_pssm.py
@@ -3,7 +3,7 @@ import numpy
 from deeprankcore.models.amino_acid import alanine
 from deeprankcore.tools.pssm import parse_pssm
 from deeprankcore.models.variant import SingleResidueVariant
-from deeprankcore.features.pssm import add_features
+from deeprankcore.features.conservation import add_features
 from deeprankcore.models.graph import build_atomic_graph
 from deeprankcore.models.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain.features import nodefeats as Nfeat
