+++ tests/feature/test_surfacearea.py
@@ -1,13 +1,16 @@
-import numpy
 from pdb2sql import pdb2sql
+import numpy
 from deeprankcore.models.amino_acid import alanine
+from deeprankcore.models.structure import Structure, Chain, Residue
 from deeprankcore.models.variant import SingleResidueVariant
-from deeprankcore.models.graph import Graph, Node
-from deeprankcore.models.structure import Chain, Residue
-from deeprankcore.features.sasa import add_features
+from deeprankcore.features.surfacearea import add_features
 from deeprankcore.models.graph import build_residue_graph, build_atomic_graph
-from deeprankcore.models.pdb import get_structure, get_surrounding_residues
+from deeprankcore.models.pdb import (
+    get_structure,
+    get_residue_contact_pairs,
+    get_surrounding_residues)
 from deeprankcore.domain.features import nodefeats
+
 
 
 def _get_residue(chain: Chain, number: int) -> Residue:
@@ -17,26 +20,84 @@ def _get_residue(chain: Chain, number: int) -> Residue:
 
     raise ValueError(f"Not found: {number}")
 
-
-def _get_residue_node(graph: Graph, number: int) -> Node:
+def _find_residue_node(graph, chain_id, residue_number):
     for node in graph.nodes:
         residue = node.id
-        if residue.number == number:
+        if residue.chain.id == chain_id and residue.number == residue_number:
             return node
 
-    raise ValueError(f"Not found: {number}")
+    raise ValueError(f"Not found: {chain_id} {residue_number}")
 
 
-def _get_atom_node(graph: Graph, residue_number: int, atom_name: str) -> Node:
+def _find_atom_node(graph, chain_id, residue_number, atom_name):
     for node in graph.nodes:
         atom = node.id
-        if atom.residue.number == residue_number and atom.name == atom_name:
+        if (
+            atom.residue.chain.id == chain_id
+            and atom.residue.number == residue_number
+            and atom.name == atom_name
+        ):
+
             return node
 
-    raise ValueError(f"Not found: {residue_number} {atom_name}")
+    raise ValueError(f"Not found: {chain_id} {residue_number} {atom_name}")
 
 
-def test_add_features_to_residues():
+def _load_pdb_structure(pdb_path: str, id_: str) -> Structure:
+    pdb = pdb2sql(pdb_path)
+    try:
+        return get_structure(pdb, id_)
+    finally:
+        pdb._close() # pylint: disable=protected-access
+
+
+def test_bsa_residue():
+    pdb_path = "tests/data/pdb/1ATN/1ATN_1w.pdb"
+
+    structure = _load_pdb_structure(pdb_path, "1ATN_1w")
+
+    residues = set([])
+    for residue1, residue2 in get_residue_contact_pairs(
+        pdb_path, structure, "A", "B", 8.5
+    ):
+        residues.add(residue1)
+        residues.add(residue2)
+
+    graph = build_residue_graph(residues, "1ATN-1w", 8.5)
+
+    add_features(pdb_path, graph)
+
+    # chain B ASP 93, at interface
+    node = _find_residue_node(graph, "B", 93)
+
+    assert node.features[nodefeats.BSA] > 0.0
+
+
+def test_bsa_atom():
+    pdb_path = "tests/data/pdb/1ATN/1ATN_1w.pdb"
+
+    structure = _load_pdb_structure(pdb_path, "1ATN_1w")
+
+    atoms = set([])
+    for residue1, residue2 in get_residue_contact_pairs(
+        pdb_path, structure, "A", "B", 8.5
+    ):
+        for atom in residue1.atoms:
+            atoms.add(atom)
+        for atom in residue2.atoms:
+            atoms.add(atom)
+    atoms = list(atoms)
+
+    graph = build_atomic_graph(atoms, "1ATN-1w", 8.5)
+
+    add_features(pdb_path, graph)
+
+    # chain B ASP 93, at interface
+    node = _find_atom_node(graph, "B", 93, "OD1")
+
+    assert node.features[nodefeats.BSA] > 0.0
+
+def test_sasa_residue():
 
     pdb_path = "tests/data/pdb/101M/101M.pdb"
 
@@ -61,17 +122,17 @@ def test_add_features_to_residues():
     )
 
     # surface residues should have large area
-    surface_residue_node = _get_residue_node(graph, 105)
+    surface_residue_node = _find_residue_node(graph, "A", 105)
     assert surface_residue_node.features[nodefeats.SASA] > 25.0
 
     # buried residues should have small area
-    buried_residue_node = _get_residue_node(graph, 72)
+    buried_residue_node = _find_residue_node(graph, "A", 72)
     assert (
         buried_residue_node.features[nodefeats.SASA] < 25.0
     ), buried_residue_node.features[nodefeats.SASA]
 
 
-def test_add_features_to_atoms():
+def test_sasa_atom():
 
     pdb_path = "tests/data/pdb/101M/101M.pdb"
 
@@ -101,11 +162,11 @@ def test_add_features_to_atoms():
     )
 
     # surface atoms should have large area
-    surface_atom_node = _get_atom_node(graph, 105, "OE2")
+    surface_atom_node = _find_atom_node(graph, "A", 105, "OE2")
     assert surface_atom_node.features[nodefeats.SASA] > 25.0
 
     # buried atoms should have small area
-    buried_atom_node = _get_atom_node(graph, 72, "CG")
+    buried_atom_node = _find_atom_node(graph, "A", 72, "CG")
     assert (
         buried_atom_node.features[nodefeats.SASA] == 0.0
     ), buried_atom_node.features[nodefeats.SASA]
